grader: stop grading when internal error occurs
Previously, the error would bubble up and become a panic. The result was
that the grader would crash and that the next time the grader starts,
the grader would pick up the same submission, run it again and probably
still crash.

This commit changes this behaviour to put a judging error on these
submissions and log an error.

Second small change is that grading will now stop once the first test
case returned without success. If one test case is wrong, we won't
bother with the rest.